Move `languageSelector` out of `/languages/` (#141174)
