+++ src/vs/editor/common/languageSelector.ts
@@ -4,7 +4,7 @@
  *--------------------------------------------------------------------------------------------*/
 
 import { IRelativePattern, match as matchGlobPattern } from 'vs/base/common/glob';
-import { URI } from 'vs/base/common/uri'; // TODO@Alex
+import { URI } from 'vs/base/common/uri';
 import { normalize } from 'vs/base/common/path';
 
 export interface LanguageFilter {
