+++ src/vs/editor/test/common/modes/languageSelector.test.ts
@@ -5,7 +5,7 @@
 
 import * as assert from 'assert';
 import { URI } from 'vs/base/common/uri';
-import { score } from 'vs/editor/common/languages/languageSelector';
+import { score } from 'vs/editor/common/languageSelector';
 
 suite('LanguageSelector', function () {
