+++ src/vs/workbench/services/languageStatus/common/languageStatusService.ts
@@ -11,7 +11,7 @@ import { compare } from 'vs/base/common/strings';
 import { ITextModel } from 'vs/editor/common/model';
 import { Command } from 'vs/editor/common/languages';
 import { LanguageFeatureRegistry } from 'vs/editor/common/languageFeatureRegistry';
-import { LanguageSelector } from 'vs/editor/common/languages/languageSelector';
+import { LanguageSelector } from 'vs/editor/common/languageSelector';
 import { IAccessibilityInformation } from 'vs/platform/accessibility/common/accessibility';
 import { registerSingleton } from 'vs/platform/instantiation/common/extensions';
 import { createDecorator } from 'vs/platform/instantiation/common/instantiation';
